sdxfv

+++ JavaFX/src/main/java/com/javafx/Modele/dao/listememoire/ListeMemoireRevueDAO.java
@@ -28,8 +28,8 @@ public class ListeMemoireRevueDAO implements RevueDAO {
 
 		this.donnees = new ArrayList<Revue>();
 
-		this.donnees.add(new Revue(1,"Random Revue","Revue super random", (float)5.7, "Visuel Random pour une revue random",new Periodicite(1)));
-		this.donnees.add(new Revue(2, "Le Quotidien", "Une revue questionnant l'aléatoire", (float)10.6,"Visuel d'un quotidien aléatoire",new Periodicite(2)));
+		this.donnees.add(new Revue(1,"Random Revue","Revue super random", 5.7, "Visuel Random pour une revue random",new Periodicite(1)));
+		this.donnees.add(new Revue(2, "Le Quotidien", "Une revue questionnant l'aléatoire", 10.6,"Visuel d'un quotidien aléatoire",new Periodicite(2)));
 	}
